Move conversation formatter
