+++ nemo_curator/services/__init__.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from .conversation_formatter import ConversationFormatter
 from .model_client import AsyncLLMClient, LLMClient
 from .nemo_client import NemoDeployClient
 from .openai_client import AsyncOpenAIClient, OpenAIClient
@@ -21,4 +22,5 @@ __all__ = [
     "AsyncOpenAIClient",
     "OpenAIClient",
     "NemoDeployClient",
+    "ConversationFormatter",
 ]
